Convert Http2Reader, Http2Writer to Kotlin (#4982)
* Rename Http2.java to .kt (and others)

* Convert Http2Reader, Http2Writer to Kotlin

+++ okhttp/src/main/java/okhttp3/internal/http2/Huffman.java
@@ -169,7 +169,7 @@ class Huffman {
     }
   }
 
-  private void addCode(int sym, int code, byte len) {
+  private void addCode(int sym, int code, int len) {
     Node terminal = new Node(sym, len);
 
     Node current = root;
